fix(security): add cwd allowlist to /api/statusline (C4)
Extract getAllowedRoots + isPathAllowed from files route into lib/allowed-roots.ts
so multiple routes can share the same allowlist logic. Add isPathAllowedAsync
convenience helper. Apply it to /api/statusline's cwd query param — previously
an attacker could probe arbitrary paths for git-repo status via ?cwd=. The files
route's behavior is unchanged (pure refactor + import).
