+++ lib/allowed-roots.ts
@@ -1,0 +1,72 @@
+import path from "path";
+import { listAllSessions } from "./session-reader";
+
+// Short-TTL cache for the allowed-roots set. Without this, every file list/read
+// request re-scans every pi session on disk just to check access. 5s is short
+// enough that newly-created cwds appear promptly; stored on globalThis so it
+// survives Next.js hot-reload.
+declare global {
+  var __piAllowedRootsCache: { roots: Set<string>; expiresAt: number } | undefined;
+}
+
+const ALLOWED_ROOTS_TTL_MS = 5_000;
+const WINDOWS_ABSOLUTE_RE = /^[a-zA-Z]:[\\/]/;
+
+export function normalizeSlashes(filePath: string): string {
+  return filePath.replace(/\\/g, "/");
+}
+
+export function isWindowsAbsolutePath(filePath: string): boolean {
+  return WINDOWS_ABSOLUTE_RE.test(filePath) || filePath.startsWith("\\\\") || filePath.startsWith("//");
+}
+
+export async function getAllowedRoots(): Promise<Set<string>> {
+  const now = Date.now();
+  const cached = globalThis.__piAllowedRootsCache;
+  if (cached && cached.expiresAt > now) return cached.roots;
+
+  const sessions = await listAllSessions();
+  const roots = new Set<string>();
+  for (const s of sessions) {
+    if (s.cwd) roots.add(s.cwd);
+  }
+  // Also allow ~/pi-cwd-* directories created by the default-cwd endpoint
+  const home = (await import("os")).homedir();
+  const { readdir } = await import("fs/promises");
+  try {
+    const names = await readdir(home);
+    for (const name of names) {
+      if (/^pi-cwd-\d{8}$/.test(name)) {
+        roots.add(path.join(home, name));
+      }
+    }
+  } catch {
+    // ignore if home is unreadable
+  }
+
+  globalThis.__piAllowedRootsCache = { roots, expiresAt: now + ALLOWED_ROOTS_TTL_MS };
+  return roots;
+}
+
+export function isPathAllowed(target: string, allowedRoots: Set<string>): boolean {
+  for (const root of allowedRoots) {
+    const useWindowsRules = isWindowsAbsolutePath(target) || isWindowsAbsolutePath(root);
+    const resolver = useWindowsRules ? path.win32 : path;
+    const sep = useWindowsRules ? "\\" : path.sep;
+    const normalized = resolver.resolve(target);
+    const normalizedRoot = resolver.resolve(root);
+    const comparable = useWindowsRules ? normalized.toLowerCase() : normalized;
+    const comparableRoot = useWindowsRules ? normalizedRoot.toLowerCase() : normalizedRoot;
+    const rootWithSep = comparableRoot.endsWith(sep) ? comparableRoot : comparableRoot + sep;
+    if (comparable === comparableRoot || comparable.startsWith(rootWithSep)) {
+      return true;
+    }
+  }
+  return false;
+}
+
+/** Convenience: get roots + check in one call. */
+export async function isPathAllowedAsync(target: string): Promise<boolean> {
+  const roots = await getAllowedRoots();
+  return isPathAllowed(target, roots);
+}
